backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.11.1
pydantic==1.10.9
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.0.0
chromadb==0.4.0
PyMuPDF==1.22.5
requests==2.31.0
httpx==0.24.0
aiofiles==23.1.0
